rename: inquiries 타입 경로 재설정 (#200)
